requirements.txt
Django==1.8.5
Unipath==1.1
argparse==1.2.1
django-braces==1.8.1
selenium==2.48.0
six==1.10.0
wsgiref==0.1.2
